Move primary and secondary editor actions into a single injectable class
